dynamically import wasm module
